refactor(ui): split chatPanelAtom into per-concern atom modules
Move the 758-line store/ui/chatPanelAtom.ts into store/ui/chatPanel/
(widthAtoms, visibilityAtoms, displayPrefsAtoms, selectionAtoms,
surfaceAtoms, miscAtoms), mirroring the workStationLayout/ convention.
chatPanelAtom.ts stays a pure re-export barrel so none of the 132 import
sites change; the width block (save timer, lastVisibleChatWidth and the
import-time getInitialChatWidth call) moved as one unit and is still
evaluated first. Pure code motion, no behaviour change.
